LC-465: renamed Stage and moved enum to stage util
- changed catch exception to catch all Exceptions instead of just runtimeException for future proofing
